Database v1.0 (Leggere descrizione)
Introdotti (migliorati per quelli già esistenti) i seguenti metodi:

1 -  Registrazione/accesso account utente:
Registrazione ✅
Login ✅
Modifica profilo utente (tabella USER) da parte degli utenti stessi ✅
−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−
2 -  Specifiche per le aste:
Aggiunta nuova asta (uno per ogni tipo) ✅
Aggiunta offerta su asta ✅
−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−
3 -  Ricerca e filtraggio aste:
Ricerca tramite keyword delle aste ✅
Ricerca tramite categoria aste ✅
Reindirizzamento al profilo del venditore di una data asta ✅(Basta utilizzare http://localhost:8080/api/user/{id} con {id} = sellerId)✅
−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−−
Metodi generali:
Notifica per partecipanti e venditore per asta fallita ✅
Notifica per partecipanti e venditore per asta conclusa con successo  ✅
Ottenere tutte le aste ✅
Ottenere tutte le aste di un tipo (1 metodo per tipo di asta) ✅
Ottenere un'asta in base al suo id ✅
Ottenere un'asta in base all'offerId ✅
Update degli attributi per ogni asta ✅
Statuscode custom per metodo getUserByUsername [SERVE A TONY PER IL FRONTEND] ✅ Statuscode 400 Bad Request
Implementare @Scheduled con timer che aggiorna i prezzi di DescendingPriceAuction ✅
Implementare @Scheduled con timer per EnglishAuction ✅
Da implementare in qualche modo l'automizzazione del passaggio di Auction.isValid da true a false quando un'asta termina ✅
Implementare metodo getNotificationByUserId ✅
Implementare metodo getAuctionBySellerId ✅
Implementare metodo Delete per ogni asta e per le notifiche ✅

+++ DietiDeals24Server/src/main/java/com/dietideals24/DietiDeals24/entity/Notification.java
@@ -16,11 +16,11 @@ public class Notification {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int notificationId;
+    private long notificationId;
     @Column
-    private int auctionId;
+    private long auctionId;
     @Column
-    private int receiverId;
+    private long receiverId;
     @Column
     private String body;
 }
